Fix terser errors in specific includeDirByFlag cases
This replaces Funnel/exclude with replace and replaces the full file content with an empty comment that is then erased by Terser

There's a standing bug in the ember-cli pipeline that reintroduces files when they are removed by Funnel with the uncompiled version,
leading to not only unwanted output but also potentially code that can't be processed by Terser.

I tried really hard to come up with a repro within the addon test suite, but it's been resisting me pretty hard.

You can find a standalone repro with a minimal new ember app at https://github.com/halfbyte/compile-repro

Just check out, do `yarn install` and then `ember build -e production`

Terser borks on a decorator in the component.

The repro uses the current commit on master to ensure compareability.